requirements.txt
streamlit
openai
awscli
boto3
Pillow
streamlit-lottie